move ExpressionCountAggregationResult implementation into own file
