+++ searchlib/src/vespa/searchlib/aggregation/aggregation.cpp
@@ -2,14 +2,8 @@
 
 #include "aggregation.h"
 
-#include "expressioncountaggregationresult.h"
-
-#include <vespa/searchlib/expression/resultvector.h>
-
 #include <vespa/searchlib/aggregation/aggregation.hpp>
 #include <vespa/vespalib/objects/visit.hpp>
-
-#include <xxhash.h>
 
 using namespace search::expression;
 
@@ -19,7 +13,6 @@ using vespalib::Deserializer;
 using vespalib::Serializer;
 
 IMPLEMENT_ABSTRACT_AGGREGATIONRESULT(AggregationResult, ExpressionNode);
-IMPLEMENT_AGGREGATIONRESULT(ExpressionCountAggregationResult, AggregationResult);
 
 AggregationResult::AggregationResult() : _expressionTree(std::make_shared<ExpressionTree>()), _tag(-1) {
 }
@@ -71,53 +64,6 @@ void AggregationResult::selectMembers(const vespalib::ObjectPredicate& predicate
                                       vespalib::ObjectOperation&       operation) {
     _expressionTree->select(predicate, operation);
 }
-
-namespace {
-// Calculates the sum of all buckets.
-template <int BucketBits, typename HashT> int calculateRank(const Sketch<BucketBits, HashT>& sketch) {
-    if (sketch.getClassId() == SparseSketch<BucketBits, HashT>::classId) {
-        return static_cast<const SparseSketch<BucketBits, HashT>&>(sketch).getSize();
-    }
-    auto normal = static_cast<const NormalSketch<BucketBits, HashT>&>(sketch);
-    int  rank = 0;
-    for (size_t i = 0; i < sketch.BUCKET_COUNT; ++i) {
-        rank += normal.bucket[i];
-    }
-    return rank;
-}
-} // namespace
-
-void ExpressionCountAggregationResult::onMerge(const AggregationResult& r) {
-    const auto& result = Identifiable::cast<const ExpressionCountAggregationResult&>(r);
-    _hll.merge(result._hll);
-    _rank.set(calculateRank(_hll.getSketch()));
-}
-void ExpressionCountAggregationResult::onAggregate(const ResultNode& result) {
-    size_t             hash = result.hash();
-    const unsigned int seed = 42;
-    hash = XXH32(&hash, sizeof(hash), seed);
-    // The rank is a maintained sum of all buckets. This should give
-    // almost the same ordering as the actual estimates.
-    _rank += _hll.aggregate(hash);
-}
-void ExpressionCountAggregationResult::onReset() {
-    _hll = HyperLogLog<PRECISION>();
-    _rank.set(0);
-}
-Serializer& ExpressionCountAggregationResult::onSerialize(Serializer& os) const {
-    AggregationResult::onSerialize(os);
-    _hll.serialize(os);
-    return os;
-}
-Deserializer& ExpressionCountAggregationResult::onDeserialize(Deserializer& is) {
-    AggregationResult::onDeserialize(is);
-    _hll.deserialize(is);
-    _rank.set(calculateRank(_hll.getSketch()));
-    return is;
-}
-
-ExpressionCountAggregationResult::ExpressionCountAggregationResult() = default;
-ExpressionCountAggregationResult::~ExpressionCountAggregationResult() = default;
 
 } // namespace search::aggregation
